Add party toggle  feature.
Adding support for toggle all markers for a specific party.

This needed a restructure of the data for people, took the opportunity
to break them out to individual files per party.

Will not by default load the two parties not currently represented in
the body. Must use toggle to see those icons.